requirements.txt
streamlit
langchain
langchain-groq
textblob
scikit-learn
numpy
pandas
python-dotenv
googletrans
transliterate